+++ lib/src/styles/colors.dart
@@ -3,14 +3,39 @@ import 'package:flutter/cupertino.dart';
 import 'macos_theme.dart';
 
 extension MacosDynamicColor on CupertinoDynamicColor {
+  /// Resolves the given [Color] by calling [resolveFrom].
+  ///
+  /// If the given color is already a concrete [Color], it will be returned as is.
+  /// If the given color is null, returns null.
+  /// If the given color is a [CupertinoDynamicColor], but the given [BuildContext]
+  /// lacks the dependencies required to the color resolution, the default trait
+  /// value will be used ([Brightness.light] platform brightness, normal contrast,
+  /// [CupertinoUserInterfaceLevelData.base] elevation level).
+  ///
+  /// See also:
+  ///
+  ///  * [resolve], which is similar to this function, but returns a
+  ///    non-nullable value, and does not allow a null `resolvable` color.
   static Color? maybeResolve(Color? resolvable, BuildContext context) {
     if (resolvable == null) return null;
     return resolve(resolvable, context);
   }
 
+  /// Resolves the given [Color] by calling [resolveFrom].
+  ///
+  /// If the given color is already a concrete [Color], it will be returned as is.
+  /// If the given color is a [CupertinoDynamicColor], but the given [BuildContext]
+  /// lacks the dependencies required to the color resolution, the default trait
+  /// value will be used ([Brightness.light] platform brightness, normal contrast,
+  /// [CupertinoUserInterfaceLevelData.base] elevation level).
+  ///
+  /// See also:
+  ///
+  ///  * [maybeResolve], which is similar to this function, but will allow a
+  ///    null `resolvable` color.
   static Color resolve(Color resolvable, BuildContext context) {
     return (resolvable is CupertinoDynamicColor)
-        ? resolvable.resolveFrom(context)
+        ? resolvable._resolveFrom(context)
         : resolvable;
   }
 
@@ -35,7 +60,7 @@ extension MacosDynamicColor on CupertinoDynamicColor {
         darkHighContrastColor != darkHighContrastElevatedColor;
   }
 
-  CupertinoDynamicColor resolveFrom(BuildContext context) {
+  CupertinoDynamicColor _resolveFrom(BuildContext context) {
     Brightness brightness = Brightness.light;
     if (this.isPlatformBrightnessDependent) {
       brightness = MacosTheme.maybeBrightnessOf(context) ?? Brightness.light;